ASSERT(!m_textCheckingRequest) on editing/spelling/spellcheck-async-remove-frame.html https://bugs.webkit.org/show_bug.cgi?id=126606
Reviewed by Wenson Hsieh.

SpellChecker's sequence numbers need to be unique between frames, because they are sent out
Source/WebCore:

to EditorClient which may be one per web view, move from a per-SpellChecker int to
TextCheckingRequestIdentifier.

Change terminology from "sequence" to "identifier" and use Optional in
TextCheckingRequestData to represent the unset state.

Tested by existing tests.

* editing/SpellChecker.cpp:
(WebCore::SpellCheckRequest::SpellCheckRequest):
(WebCore::SpellCheckRequest::didSucceed):
(WebCore::SpellCheckRequest::didCancel):
(WebCore::SpellCheckRequest::setCheckerAndIdentifier):
(WebCore::SpellChecker::SpellChecker):
(WebCore::SpellChecker::requestCheckingFor):
(WebCore::SpellChecker::didCheck):
(WebCore::SpellChecker::didCheckSucceed):
(WebCore::SpellChecker::didCheckCancel):
(WebCore::SpellCheckRequest::setCheckerAndSequence): Deleted.
* editing/SpellChecker.h:
(WebCore::SpellChecker::lastRequestIdentifier const):
(WebCore::SpellChecker::lastProcessedIdentifier const):
(WebCore::SpellChecker::lastRequestSequence const): Deleted.
(WebCore::SpellChecker::lastProcessedSequence const): Deleted.
* platform/text/TextChecking.h:
(WebCore::TextCheckingRequestData::TextCheckingRequestData):
(WebCore::TextCheckingRequestData::identifier const):
(WebCore::TextCheckingRequestData::sequence const): Deleted.
* testing/Internals.cpp:
(WebCore::Internals::lastSpellCheckRequestSequence):
(WebCore::Internals::lastSpellCheckProcessedSequence):
* testing/Internals.h:
* testing/Internals.idl:

Source/WebKit:

to EditorClient which may be one per web view, so use a static to hold the current
sequence number.

Also convert from int to TextCheckingSequence to make them easier to find, and use
Markable and Optional in TextCheckingRequestData to represent the unset state.

* Shared/WebCoreArgumentCoders.cpp:
(IPC::ArgumentCoder<TextCheckingRequestData>::encode):
(IPC::ArgumentCoder<TextCheckingRequestData>::decode):

Source/WebKitLegacy/mac:

to EditorClient which may be one per web view, so use a static to hold the current
sequence number.

Also convert from int to TextCheckingSequence to make them easier to find, and use
Markable and Optional in TextCheckingRequestData to represent the unset state.

* WebCoreSupport/WebEditorClient.h:
* WebCoreSupport/WebEditorClient.mm:
(-[WebEditorSpellCheckResponder initWithClient:identifier:results:]):
(-[WebEditorSpellCheckResponder perform]):
(WebEditorClient::didCheckSucceed):
(WebEditorClient::requestCheckingOfString):
(-[WebEditorSpellCheckResponder initWithClient:sequence:results:]): Deleted.

Canonical link: https://commits.webkit.org/221773@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258182 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/testing/Internals.h
@@ -304,8 +304,8 @@ public:
 
     ExceptionOr<void> setDelegatesScrolling(bool enabled);
 
-    ExceptionOr<int> lastSpellCheckRequestSequence();
-    ExceptionOr<int> lastSpellCheckProcessedSequence();
+    ExceptionOr<uint64_t> lastSpellCheckRequestSequence();
+    ExceptionOr<uint64_t> lastSpellCheckProcessedSequence();
 
     Vector<String> userPreferredLanguages() const;
     void setUserPreferredLanguages(const Vector<String>&);
